+++ app/src/main/java/com/numero/material_gallery/AppModule.kt
@@ -3,8 +3,8 @@ package com.numero.material_gallery
 import android.content.Context
 import com.google.android.play.core.appupdate.AppUpdateManager
 import com.google.android.play.core.appupdate.AppUpdateManagerFactory
-import com.numero.material_gallery.repository.ConfigRepository
-import com.numero.material_gallery.repository.ConfigRepositoryImpl
+import com.numero.material_gallery.core.repository.ConfigRepository
+import com.numero.material_gallery.core.repository.ConfigRepositoryImpl
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
