+++ app/src/main/java/com/numero/material_gallery/MaterialGalleryApplication.kt
@@ -1,7 +1,7 @@
 package com.numero.material_gallery
 
 import android.app.Application
-import com.numero.material_gallery.repository.ConfigRepository
+import com.numero.material_gallery.core.repository.ConfigRepository
 import dagger.hilt.android.HiltAndroidApp
 import javax.inject.Inject
