+++ core/build.gradle
@@ -40,6 +40,7 @@ dependencies {
     implementation libs.androidx.appcompat
     implementation libs.androidx.core
     implementation libs.androidx.lifecycle.runtime
+    implementation libs.androidx.preference
 
     implementation libs.material
